[Hyperdrive Trading] Chart component
